+++ files/gnome/dconf.nix
@@ -1,11 +1,17 @@
-{ config, lib, ... }:
+{ config, lib, files, ... }:
 with lib.hm.gvariant;
 let homeDir = config.home.homeDirectory;
 in {
-  # GTK+ Bookmarks
-  home.file.".config/gtk-3.0/bookmarks".text = lib.mkBefore ''
-    file://${homeDir}/Documents/TBD TBD
-  '';
+  # Home Directory
+  home.file = {
+    # Wallpapers
+    ".local/share/backgrounds".source = files.images.wallpapers;
+
+    # GTK+ Bookmarks
+    ".config/gtk-3.0/bookmarks".text = lib.mkBefore ''
+      file://${homeDir}/Documents/TBD TBD
+    '';
+  };
 
   ## Dconf Keys ##
   # Generated via dconf2nix: https://github.com/gvolpe/dconf2nix