+++ modules/user/security.nix
@@ -8,7 +8,6 @@ in rec {
   config = {
     # Passwords
     sops.secrets = util.map.secrets ./passwords true;
-    users.extraUsers.root.passwordFile = secrets."root.secret".path;
     users.users = mapAttrs (name: _: {
       passwordFile =
         mkIf (!users."${name}".minimal) secrets."${name}.secret".path;